requirements.txt
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.1.0
Beaker==1.12.1
beautifulsoup4==4.13.5
blivet==3.12.1
blivet-gui==2.6.0
certifi==2025.7.9
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cupshelpers==1.0
dasbus==1.7
dbus_next==0.2.3
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnf==4.23.0
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
FVS==0.3.4
gunicorn==23.0.0
humanize==4.12.0
icoextract==0.1.5
idna==3.10
iso639==0.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.69
louis==3.33.0
Mako==1.2.3
Markdown==3.7
meson==1.7.2
mutagen==1.47.0
nftables==0.1
olefile==0.47
packaging==24.2
Paste==3.10.1
pathvalidate==3.2.3
patool==4.0.0
pefile==2024.8.26
pexpect==4.9.0
pid==2.2.3
ply==3.11
productmd==1.45
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.20
pyenchant==3.2.2
Pygments==2.18.0
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart==3.62
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pyudev==0.24.3
pyxdg==0.27
pyynl @ file:///builddir/build/BUILD/kernel-6.16.5-build/kernel-6.16.5/linux-6.16.5-200.fc42.x86_64/tools/net/ynl
referencing==0.36.2
reportlab==4.4.1
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setuptools==74.1.3
simpleline==1.9.0
six==1.17.0
sos==4.10.0
soupsieve==2.7
sqlparse==0.5.3
Tempita==0.5.2
typing_extensions==4.12.2
urllib3==2.3.0
vkbasalt-cli==3.1.1.post2
whitenoise==6.10.0
yt-dlp==2025.9.5
